check timestamp of images
By default, Imlib2 stores an image with a given filename on its cache
and, if it loads a image with the same filename, it will just get the
image already on the cache.  But if the file changes, it will get the
old version (the version on the cache).   Now, xnotify will check the
timestamp of an image on disk and compare it with the image in its
cache and reload it if the timestamps differ.