several changes
I adopted the ctrlfnt mechanism I have been using in other projects.
It handles both Xft and XLFD fonts.

I added the -r command-line option to use X Properties in addition of
the standard input to read notification strings from.  This fixes #37

I added the XNotify.opacity X resource to specify the window opacity.
This implements transparency as #32

There is much to be tested to check for bugs that could have been
introduced.